requirements.txt
av
numpy
opencv-python
Pillow
pygame
pyside6
scipy
tqdm